시간표 조회, 수정 및 출석 체크시 현재 로그인된 유저와 동일한지 검증 후 기능이 수행되는 절차 추가

+++ src/main/java/me/asite/timetable/TimetableController.java
@@ -3,6 +3,8 @@ package me.asite.timetable;
 import lombok.RequiredArgsConstructor;
 import me.asite.attendance.AttendanceCheckRequestDto;
 import me.asite.exception.CannotFindByIDException;
+import me.asite.student.CurrentStudent;
+import me.asite.student.Student;
 import org.springframework.hateoas.Link;
 import org.springframework.hateoas.MediaTypes;
 import org.springframework.hateoas.Resources;
@@ -25,9 +27,10 @@ public class TimetableController {
 
     @PostMapping("/add")
     public ResponseEntity addTimetable(@RequestBody @Valid TimetableAddRequestDto dto,
+                                       @CurrentStudent Student student,
                                        Errors errors) {
 
-        if (errors.hasErrors()) {
+        if (errors.hasErrors() || !student.getId().equals(dto.getStudentId())) {
             return ResponseEntity.badRequest().build();
         }
 
@@ -38,16 +41,19 @@ public class TimetableController {
         return ResponseEntity.ok(timetableResource);
     }
 
-    @GetMapping("/queryByStudentId/{id}")
-    public ResponseEntity query_TimetableByStudentId(@PathVariable("id") Long studentId) {
-
+    @GetMapping("/queryTimetable")
+    public ResponseEntity query_TimetableByStudentId(@RequestParam Long studentId,
+                                                     @CurrentStudent Student student) {
         try {
 
+            if (!student.getId().equals(studentId)) {
+                return ResponseEntity.badRequest().build();
+            }
+
             List<Timetable> findTimetables = timetableService.findAllWithCourseByStudentId(studentId);
-            List<TimetableResource> collect = findTimetables.stream()
-                    .map(TimetableResource::new)
-                    .collect(toList());
-            Resources<TimetableResource> timetableResources = new Resources<>(collect);
+
+            Resources<TimetableResource> timetableResources = getTimetableResources(findTimetables);
+
             timetableResources.add(linkTo(TimetableController.class).slash("queryByStudentId").slash(studentId).withSelfRel());
             timetableResources.add(new Link("/docs/index.html#resources-query-timetable").withRel("profile"));
             return ResponseEntity.ok(timetableResources);
@@ -59,13 +65,18 @@ public class TimetableController {
 
     @PutMapping("/{id}")
     public ResponseEntity attendanceCheck(@PathVariable("id") Long timetableId,
+                                          @RequestParam Long studentId,
+                                          @CurrentStudent Student student,
                                           @RequestBody @Valid AttendanceCheckRequestDto dto,
                                           Errors errors) {
-        if (errors.hasErrors()) {
+
+        if (errors.hasErrors() || !student.getId().equals(studentId)) {
             return ResponseEntity.badRequest().build();
         }
 
         Timetable checkedTimetable = timetableService.attendanceCheck(timetableId, dto);
+
+
         TimetableResource timetableResource = new TimetableResource(checkedTimetable);
         timetableResource.add(new Link("/docs/index.html#resources-attendance-check").withRel("profile"));
 
@@ -73,11 +84,23 @@ public class TimetableController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity deleteTimetable(@PathVariable("id") Long timetableId) {
+    public ResponseEntity deleteTimetable(@PathVariable("id") Long timetableId,
+                                          @RequestParam Long studentId,
+                                          @CurrentStudent Student student) {
+
+        if (!student.getId().equals(studentId)) {
+            return ResponseEntity.badRequest().build();
+        }
+
         timetableService.deleteTimetable(timetableId);
-        //TODO 현재 유저가 시간표의 주인인지 확인하는 검증 필요
-        Link link = new Link("/docs/index.html#resources-attendance-check");
         return ResponseEntity.ok().build();
     }
 
+
+    private Resources<TimetableResource> getTimetableResources(List<Timetable> findTimetables) {
+        List<TimetableResource> collect = findTimetables.stream()
+                .map(TimetableResource::new)
+                .collect(toList());
+        return new Resources<>(collect);
+    }
 }
